Refactor: rename prospector type and other references to input
This refactor rename the prospector type to the input type, this allow
this project to be more aligned with Logstash naming convention and also remove
some of the last naming legacy of the `logstash fowarder`. The input
name is also more appropriate for the UDP and the Redis code.

The prospectors are now moved to the `input` folder. Backward
compatibility is keep by using type aliasing over the older `prospector`
types. Logs statements were also changed to reflect this refactor.

Currently all the code and YAML are still using the *prospector(s)* keys,
but other PRs will move the usage to the inputs. If the `input(s)` are
present we will use them instead of the *prospectors* key.